Add support for redirect_uri option
ChannelAdvisor only allows a single Redirect URI, and it must exactly match the
value provided when the application was registered in the CA Developer Console.

The default behavior for the OAuth2 strategy is to build the URI based on the
current host. This option allows you to override the value passed back to CA,
regardless of the current host, so that you can make sure to match the registered
value.